+++ Source/EventFlow.Hangfire/Integration/EventFlowHangfireOptions.cs
@@ -21,11 +21,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace EventFlow.Hangfire.Integration
 {
@@ -40,7 +36,7 @@ namespace EventFlow.Hangfire.Integration
 
         public IEventFlowHangfireOptions UseQueueName(string queueName)
         {
-            _eventFlowOptions.RegisterServices(sr => sr.Register<IQueueNameProvider>(r => new QueueNameProvider(queueName)));
+            _eventFlowOptions.ServiceCollection.TryAddTransient<IQueueNameProvider>(r => new QueueNameProvider(queueName));
             return this;
         }
     }