+++ Source/EventFlow.Hangfire/Integration/IEventFlowHangfireOptions.cs
@@ -21,12 +21,6 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace EventFlow.Hangfire.Integration
 {
     public interface IEventFlowHangfireOptions